refactor: rename src/core/ to src/logging/
