+++ main/src/Adapter/Adapter.cpp
@@ -1,6 +1,6 @@
 #include "Adapter.h"
 #include "src/Mesh/Mesh.h" // for full definition of mesh_message
-#include "src/core/Logger.h"
+#include "src/logging/Logger.h"
 #include "src/error/Error.h"
 #include "src/Adapter/AdapterFactory.h"
 #include "src/Adapter/Serial_Adapter/Serial_Adapter.h"
