+++ main/src/Adapter/Serial_Adapter/Serial_Adapter.cpp
@@ -1,6 +1,6 @@
 #include "Serial_Adapter.h"
 #include "src/Adapter/AdapterFactory.h"
-#include "src/core/Logger.h"
+#include "src/logging/Logger.h"
 #include "src/error/Error.h"
 #include <esp_wifi.h>
 #include "src/Mesh/Mesh.h"
